+++ src/main/scala/com/databricks/labs/mosaic/core/types/package.scala
@@ -7,6 +7,7 @@ import org.apache.spark.sql.types._
   * for type inference over geometry columns.
   */
 package object types {
+
     val HexType: DataType = new HexType()
     val JSONType: DataType = new JSONType()
     // Note InternalGeometryType depends on InternalCoordType
@@ -15,4 +16,5 @@ package object types {
     val BoundaryType: ArrayType = ArrayType(ArrayType(InternalCoordType))
     val HolesType: ArrayType = ArrayType(ArrayType(ArrayType(InternalCoordType)))
     val InternalGeometryType: DataType = new InternalGeometryType()
+
 }